fix: use Float64 for DeduplicatedReads in featureDump schema
EM-based quantification methods (e.g. parsimony-EM) produce fractional
deduplicated read counts. The hardcoded Int64 type caused a parse failure
when converting to AnnData.

Fixes COMBINE-lab/simpleaf#194

+++ af-anndata/src/lib.rs
@@ -545,7 +545,7 @@ pub fn convert_csr_to_anndata<P: AsRef<Path>>(root_path: P, output_path: P) -> a
     let feat_dump_schema = Arc::new(Schema::from_iter([
         Field::new("CorrectedReads".into(), DataType::Int64),
         Field::new("MappedReads".into(), DataType::Int64),
-        Field::new("DeduplicatedReads".into(), DataType::Int64),
+        Field::new("DeduplicatedReads".into(), DataType::Float64),
         Field::new("MappingRate".into(), DataType::Float64),
         Field::new("DedupRate".into(), DataType::Float64),
         Field::new("MeanByMax".into(), DataType::Float64),
